делает файл валидации
